django/requirements.txt
Django==1.9.3
django-webpack-loader==0.5.0
djangorestframework==3.6.4
pytz==2017.2
